feature: add tests
(cherry picked from commit dcd575383a62b673e017aa91348d9021fa840781)

+++ src/main/java/com/skyeng/SEPost/dto/PostOfficeDto.java
@@ -10,7 +10,7 @@ import lombok.experimental.FieldDefaults;
 @NoArgsConstructor
 public class PostOfficeDto {
 
-    Long index;
+    Integer index;
     String title;
     String address;
 }
