+++ src/main/java/com/skyeng/SEPost/repository/PostOfficeRepository.java
@@ -3,5 +3,5 @@ package com.skyeng.SEPost.repository;
 import com.skyeng.SEPost.entity.PostOffice;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PostOfficeRepository extends JpaRepository<PostOffice, Long> {
+public interface PostOfficeRepository extends JpaRepository<PostOffice, Integer> {
 }
